Split BSShader export into BSShaderProperty block and BSTexture classes sharing across nitextureproperty classes.
