+++ io_scene_nif/modules/nif_export/animation/shader.py
@@ -1,0 +1,57 @@
+"""This script contains classes to help export shader based animations."""
+
+# ***** BEGIN LICENSE BLOCK *****
+#
+# Copyright © 2020, NIF File Format Library and Tools contributors.
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions
+# are met:
+#
+#    * Redistributions of source code must retain the above copyright
+#      notice, this list of conditions and the following disclaimer.
+#
+#    * Redistributions in binary form must reproduce the above
+#      copyright notice, this list of conditions and the following
+#      disclaimer in the documentation and/or other materials provided
+#      with the distribution.
+#
+#    * Neither the name of the NIF File Format Library and Tools
+#      project nor the names of its contributors may be used to endorse
+#      or promote products derived from this software without specific
+#      prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+# COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+#
+# ***** END LICENSE BLOCK *****
+
+from io_scene_nif.modules.nif_export.animation import Animation
+from io_scene_nif.modules.nif_export.block_registry import block_store
+
+
+class ShaderAnimation(Animation):
+
+    def __init__(self):
+        super().__init__()
+
+    def export_bs_effect_shader(self, b_mat, bsshader, b_slot):
+
+        # TODO [shader][animation] Do some form of check to ensure that we actually have data
+        effect_control = block_store.create_block("BSEffectShaderPropertyFloatController", bsshader)
+        effect_control.flags = b_mat.niftools_alpha.textureflag
+        effect_control.frequency = b_slot.texture.image.fps
+        effect_control.start_time = b_slot.texture.image.frame_start
+        effect_control.stop_time = b_slot.texture.image.frame_end
+        bsshader.add_controller(effect_control)
