+++ io_scene_nif/modules/nif_export/property/texture/__init__.py
@@ -40,7 +40,6 @@
 from pyffi.formats.nif import NifFormat
 
 from io_scene_nif.modules.nif_export.animation.texture import TextureAnimation
-from io_scene_nif.modules.nif_export.block_registry import block_store
 from io_scene_nif.modules.nif_export.property import texture
 from io_scene_nif.modules.nif_export.property.texture.writer import TextureWriter
 from io_scene_nif.utils import util_math
@@ -65,7 +64,7 @@ def get_used_textslots(b_mat):
     return used_slots
 
 
-class Texture:
+class TextureSlotManager:
     # Default ordering of Extra data blocks for different games
     USED_EXTRA_SHADER_TEXTURES = {
         'SID_MEIER_S_RAILROADS': (3, 0, 4, 1, 5, 2),
@@ -105,153 +104,6 @@ class Texture:
 
         NifLog.warn("Unsupported blend type ({0}) in material, using apply mode APPLY_MODULATE".format(b_blend_type))
         return NifFormat.ApplyMode.APPLY_MODULATE
-
-    def export_texturing_property(self, flags=0x0001, applymode=None, b_mat=None):
-        """Export texturing property."""
-
-        self.determine_texture_types(b_mat)
-
-        texprop = NifFormat.NiTexturingProperty()
-
-        texprop.flags = flags
-        texprop.apply_mode = applymode
-        texprop.texture_count = 7
-
-        self.export_texture_shader_effect(texprop)
-        self.export_nitextureprop_tex_descs(texprop)
-
-        # search for duplicate
-        for n_block in block_store.block_to_obj:
-            if isinstance(n_block, NifFormat.NiTexturingProperty) and n_block.get_hash() == texprop.get_hash():
-                return n_block
-
-        # no texturing property with given settings found, so use and register
-        # the new one
-        return texprop
-
-    def export_nitextureprop_tex_descs(self, texprop):
-
-        if self.base_mtex:
-            texprop.has_base_texture = True
-            self.texture_writer.export_tex_desc(texdesc=texprop.base_texture,
-                                                uvlayers=self.dict_mesh_uvlayers,
-                                                b_mat_texslot=self.base_mtex)
-            # check for texture flip definition
-            try:
-                fliptxt = Blender.Text.Get(basemtex.texture.name)
-            except NameError:
-                pass
-            else:
-                # texture slot 0 = base
-                # TODO [animation] FIXME Heirarchy
-                # self.texture_anim.export_flip_controller(fliptxt, self.base_mtex.texture, texprop, 0)
-                pass
-
-        if self.glow_mtex:
-            texprop.has_glow_texture = True
-            self.texture_writer.export_tex_desc(texdesc=texprop.glow_texture,
-                                                uvlayers=self.dict_mesh_uvlayers,
-                                                b_mat_texslot=self.glow_mtex)
-
-        if self.bump_mtex:
-            if NifOp.props.game not in self.USED_EXTRA_SHADER_TEXTURES:
-                texprop.has_bump_map_texture = True
-                self.texture_writer.export_tex_desc(texdesc=texprop.bump_map_texture,
-                                                    uvlayers=self.dict_mesh_uvlayers,
-                                                    b_mat_texslot=self.bump_mtex)
-                texprop.bump_map_luma_scale = 1.0
-                texprop.bump_map_luma_offset = 0.0
-                texprop.bump_map_matrix.m_11 = 1.0
-                texprop.bump_map_matrix.m_12 = 0.0
-                texprop.bump_map_matrix.m_21 = 0.0
-                texprop.bump_map_matrix.m_22 = 1.0
-
-        if self.normal_mtex:
-            shadertexdesc = texprop.shader_textures[1]
-            shadertexdesc.is_used = True
-            shadertexdesc.texture_data.source = TextureWriter.export_source_texture(n_texture=self.normal_mtex.texture)
-
-        if self.gloss_mtex:
-            if NifOp.props.game not in self.USED_EXTRA_SHADER_TEXTURES:
-                texprop.has_gloss_texture = True
-                self.texture_writer.export_tex_desc(texdesc=texprop.gloss_texture,
-                                                    uvlayers=self.dict_mesh_uvlayers,
-                                                    b_mat_texslot=self.gloss_mtex)
-            else:
-                shadertexdesc = texprop.shader_textures[2]
-                shadertexdesc.is_used = True
-                shadertexdesc.texture_data.source = TextureWriter.export_source_texture(n_texture=self.gloss_mtex.texture)
-
-        if self.dark_mtex:
-            texprop.has_dark_texture = True
-            self.texture_writer.export_tex_desc(texdesc=texprop.dark_texture,
-                                                uvlayers=self.dict_mesh_uvlayers,
-                                                b_mat_texslot=self.dark_mtex)
-
-        if self.detail_mtex:
-            texprop.has_detail_texture = True
-            self.texture_writer.export_tex_desc(texdesc=texprop.detail_texture,
-                                                uvlayers=self.dict_mesh_uvlayers,
-                                                b_mat_texslot=self.detail_mtex)
-
-        if self.ref_mtex:
-            if NifOp.props.game not in self.USED_EXTRA_SHADER_TEXTURES:
-                NifLog.warn("Cannot export reflection texture for this game.")
-                # tex_prop.hasRefTexture = True
-                # self.export_tex_desc(texdesc=tex_prop.refTexture, uvlayers=uvlayers, mtex=refmtex)
-            else:
-                shadertexdesc = texprop.shader_textures[3]
-                shadertexdesc.is_used = True
-                shadertexdesc.texture_data.source = TextureWriter.export_source_texture(n_texture=self.ref_mtex.texture)
-
-    def export_texture_shader_effect(self, tex_prop):
-        # export extra shader textures
-        if NifOp.props.game == 'SID_MEIER_S_RAILROADS':
-            # sid meier's railroads:
-            # some textures end up in the shader texture list there are 5 slots available, so set them up
-            tex_prop.num_shader_textures = 5
-            tex_prop.shader_textures.update_size()
-            for mapindex, shadertexdesc in enumerate(tex_prop.shader_textures):
-                # set default values
-                shadertexdesc.is_used = False
-                shadertexdesc.map_index = mapindex
-
-            # some texture slots required by the engine
-            shadertexdesc_envmap = tex_prop.shader_textures[0]
-            shadertexdesc_envmap.is_used = True
-            shadertexdesc_envmap.texture_data.source = TextureWriter.export_source_texture(filename="RRT_Engine_Env_map.dds")
-
-            shadertexdesc_cubelightmap = tex_prop.shader_textures[4]
-            shadertexdesc_cubelightmap.is_used = True
-            shadertexdesc_cubelightmap.texture_data.source = TextureWriter.export_source_texture(filename="RRT_Cube_Light_map_128.dds")
-
-        elif NifOp.props.game == 'CIVILIZATION_IV':
-            # some textures end up in the shader texture list there are 4 slots available, so set them up
-            tex_prop.num_shader_textures = 4
-            tex_prop.shader_textures.update_size()
-            for mapindex, shadertexdesc in enumerate(tex_prop.shader_textures):
-                # set default values
-                shadertexdesc.is_used = False
-                shadertexdesc.map_index = mapindex
-
-    def export_texture_effect(self, b_mat_texslot=None):
-        """Export a texture effect block from material texture mtex (MTex, not Texture)."""
-        texeff = NifFormat.NiTextureEffect()
-        texeff.flags = 4
-        texeff.rotation.set_identity()
-        texeff.scale = 1.0
-        texeff.model_projection_matrix.set_identity()
-        texeff.texture_filtering = NifFormat.TexFilterMode.FILTER_TRILERP
-        texeff.texture_clamping = NifFormat.TexClampMode.WRAP_S_WRAP_T
-        texeff.texture_type = NifFormat.EffectType.EFFECT_ENVIRONMENT_MAP
-        texeff.coordinate_generation_type = NifFormat.CoordGenType.CG_SPHERE_MAP
-        if b_mat_texslot:
-            texeff.source_texture = TextureWriter.export_source_texture(b_mat_texslot.texture)
-            if NifOp.props.game == 'MORROWIND':
-                texeff.num_affected_node_list_pointers += 1
-                texeff.affected_node_list_pointers.update_size()
-        texeff.unknown_vector.x = 1.0
-        return block_store.register_block(texeff)
 
     def add_shader_integer_extra_datas(self, trishape):
         """Add extra data blocks for shader indices."""
@@ -437,7 +289,7 @@ def has_diffuse_textures(self, b_mat):
         return self.diffusetextures
 
     for b_mat_texslot in texture.get_used_textslots(b_mat):
-        if b_mat_texslot.use and b_mat_texslot.use_map_color_diffuse:
+        if b_mat_texslot.use_map_color_diffuse:
             self.diffusetextures.append(b_mat_texslot)
     return self.diffusetextures
 
@@ -447,7 +299,7 @@ def has_glow_textures(self, b_mat):
         return self.glowtextures
 
     for b_mat_texslot in texture.get_used_textslots(b_mat):
-        if b_mat_texslot.use and b_mat_texslot.use_map_emit:
+        if b_mat_texslot.use_map_emit:
             self.glowtextures.append(b_mat_texslot)
     return self.glowtextures
 
@@ -457,9 +309,8 @@ def has_bumpmap_textures(self, b_mat):
         return self.bumpmaptextures
 
     for b_mat_texslot in texture.get_used_textslots(b_mat):
-        if b_mat_texslot.use:
-            if b_mat_texslot.texture.use_normal_map is False and b_mat_texslot.use_map_color_diffuse is False:
-                self.bumpmaptextures.append(b_mat_texslot)
+        if b_mat_texslot.texture.use_normal_map is False and b_mat_texslot.use_map_color_diffuse is False:
+            self.bumpmaptextures.append(b_mat_texslot)
     return self.bumpmaptextures
 
 
@@ -468,7 +319,7 @@ def has_gloss_textures(self, b_mat):
         return self.glosstextures
 
     for b_mat_texslot in texture.get_used_textslots(b_mat):
-        if b_mat_texslot.use and b_mat_texslot.use_map_color_spec:
+        if b_mat_texslot.use_map_color_spec:
             self.glosstextures.append(b_mat_texslot)
     return self.glosstextures
 
@@ -478,7 +329,6 @@ def has_normalmap_textures(self, b_mat):
         return self.normalmaptextures
 
     for b_mat_texslot in texture.get_used_textslots(b_mat):
-        if b_mat_texslot.use:
-            if b_mat_texslot.use_map_color_diffuse is False and b_mat_texslot.texture.use_normal_map and b_mat_texslot.use_map_normal:
-                self.normalmaptextures.append(b_mat_texslot)
+        if b_mat_texslot.use_map_color_diffuse is False and b_mat_texslot.texture.use_normal_map and b_mat_texslot.use_map_normal:
+            self.normalmaptextures.append(b_mat_texslot)
     return self.normalmaptextures
