+++ io_scene_nif/modules/nif_import/property/texture/types/nitextureprop.py
@@ -1,3 +1,42 @@
+"""This modules contains helper methods to import nitexturepropery based texture."""
+
+# ***** BEGIN LICENSE BLOCK *****
+#
+# Copyright © 2020, NIF File Format Library and Tools contributors.
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions
+# are met:
+#
+# * Redistributions of source code must retain the above copyright
+#   notice, this list of conditions and the following disclaimer.
+#
+# * Redistributions in binary form must reproduce the above
+#   copyright notice, this list of conditions and the following
+#   disclaimer in the documentation and/or other materials provided
+#   with the distribution.
+#
+# * Neither the name of the NIF File Format Library and Tools
+#   project nor the names of its contributors may be used to endorse
+#   or promote products derived from this software without specific
+#   prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+# COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+#
+# ***** END LICENSE BLOCK *****
+
 from io_scene_nif.modules.nif_import.property.texture import TextureSlotManager
 from io_scene_nif.utils.util_logging import NifLog
 
@@ -94,9 +133,9 @@ class NiTextureProp(TextureSlotManager):
         # 'unknown_2_texture',
         # 'unknown_2_float',
 
-        self.import_apply_mode(n_texture_desc, b_texture)
+        self._import_apply_mode(n_texture_desc, b_texture)
 
-    def import_apply_mode(self, n_texture_desc, b_texture):
+    def _import_apply_mode(self, n_texture_desc, b_texture):
         # Blend mode
         if hasattr(n_texture_desc, "apply_mode"):
             b_texture.blend_type = self.get_b_blend_type_from_n_apply_mode(n_texture_desc.apply_mode)
